fix: add nosec comments for gosec G115 with proper overflow checks

+++ api/traffic_control.go
@@ -469,7 +469,7 @@ func (tc *TrafficController) Apply() error {
 			if j > 65435 { // Prevent overflow when adding 100 (65535 - 100)
 				return fmt.Errorf("too many filters: would overflow uint16")
 			}
-			priority := uint16(100 + j) // Start filter priorities at 100
+			priority := uint16(100) + uint16(j) // Start filter priorities at 100 #nosec G115 -- overflow check performed above
 			protocol := "ip"
 			flowID := classID
